+++ packages/manager/src/features/Account/accountLandingLazyRoute.ts
@@ -1,0 +1,7 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { AccountLanding } from 'src/features/Account/AccountLanding';
+
+export const accountLandingLazyRoute = createLazyRoute('/account')({
+  component: AccountLanding,
+});
