+++ packages/manager/src/features/Betas/betaSignupLazyRoute.ts
@@ -1,0 +1,7 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { BetaSignup } from './BetaSignup';
+
+export const betaSignupLazyRoute = createLazyRoute('/betas/signup/$betaId')({
+  component: BetaSignup,
+});
