+++ packages/manager/src/features/Betas/betasLandingLazyRoute.ts
@@ -1,0 +1,7 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { BetasLanding } from './BetasLanding';
+
+export const betasLandingLazyRoute = createLazyRoute('/betas')({
+  component: BetasLanding,
+});
