+++ packages/manager/src/features/CloudPulse/Alerts/AlertsDetail/cloudPulseAlertsDefinitionsDetailLazyRoute.ts
@@ -1,0 +1,9 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { AlertDetail } from './AlertDetail';
+
+export const cloudPulseAlertsDefinitionsDetailLazyRoute = createLazyRoute(
+  '/alerts/definitions/detail/$serviceType/$alertId'
+)({
+  component: AlertDetail,
+});
