+++ packages/manager/src/features/CloudPulse/Alerts/AlertsListing/cloudPulseAlertsLandingLazyRoute.ts
@@ -1,0 +1,9 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { AlertListing } from './AlertListing';
+
+export const cloudPulseAlertsLandingLazyRoute = createLazyRoute(
+  '/alerts/definitions'
+)({
+  component: AlertListing,
+});
