+++ packages/manager/src/features/CloudPulse/Alerts/CreateAlert/cloudPulseAlertsCreateLazyRoute.ts
@@ -1,0 +1,9 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { CreateAlertDefinition } from './CreateAlertDefinition';
+
+export const cloudPulseAlertsCreateLazyRoute = createLazyRoute(
+  '/alerts/definitions/create'
+)({
+  component: CreateAlertDefinition,
+});
