+++ packages/manager/src/features/DataStream/Streams/StreamCreate/streamCreateLazyRoute.ts
@@ -1,0 +1,9 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { StreamCreate } from 'src/features/DataStream/Streams/StreamCreate/StreamCreate';
+
+export const streamCreateLazyRoute = createLazyRoute(
+  '/datastream/streams/create'
+)({
+  component: StreamCreate,
+});
