+++ packages/manager/src/features/DataStream/dataStreamLandingLazyRoute.ts
@@ -1,0 +1,7 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { DataStreamLanding } from 'src/features/DataStream/DataStreamLanding';
+
+export const dataStreamLandingLazyRoute = createLazyRoute('/datastream')({
+  component: DataStreamLanding,
+});
