+++ packages/manager/src/features/Databases/DatabaseCreate/databaseCreateLazyRoute.ts
@@ -1,0 +1,7 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { DatabaseCreate } from './DatabaseCreate';
+
+export const databaseCreateLazyRoute = createLazyRoute('/databases/create')({
+  component: DatabaseCreate,
+});
