+++ packages/manager/src/features/Databases/DatabaseDetail/databaseDetailLazyRoute.ts
@@ -1,0 +1,9 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { DatabaseDetail } from 'src/features/Databases/DatabaseDetail';
+
+export const databaseDetailLazyRoute = createLazyRoute(
+  '/databases/$engine/$databaseId'
+)({
+  component: DatabaseDetail,
+});
