+++ packages/manager/src/features/Databases/DatabaseLanding/databaseLandingLazyRoute.ts
@@ -1,0 +1,7 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { DatabaseLanding } from 'src/features/Databases/DatabaseLanding/DatabaseLanding';
+
+export const databaseLandingLazyRoute = createLazyRoute('/databases')({
+  component: DatabaseLanding,
+});
