+++ packages/manager/src/routes/databases/index.ts
@@ -20,14 +20,18 @@ const databasesIndexRoute = createRoute({
   getParentRoute: () => databasesRoute,
   path: '/',
 }).lazy(() =>
-  import('./databasesLazyRoutes').then((m) => m.databaseLandingLazyRoute)
+  import(
+    'src/features/Databases/DatabaseLanding/databaseLandingLazyRoute'
+  ).then((m) => m.databaseLandingLazyRoute)
 );
 
 const databasesCreateRoute = createRoute({
   getParentRoute: () => databasesRoute,
   path: 'create',
 }).lazy(() =>
-  import('./databasesLazyRoutes').then((m) => m.databaseCreateLazyRoute)
+  import('src/features/Databases/DatabaseCreate/databaseCreateLazyRoute').then(
+    (m) => m.databaseCreateLazyRoute
+  )
 );
 
 const databasesDetailRoute = createRoute({
@@ -47,56 +51,72 @@ const databasesDetailRoute = createRoute({
   },
   path: '$engine/$databaseId',
 }).lazy(() =>
-  import('./databasesLazyRoutes').then((m) => m.databaseDetailLazyRoute)
+  import('src/features/Databases/DatabaseDetail/databaseDetailLazyRoute').then(
+    (m) => m.databaseDetailLazyRoute
+  )
 );
 
 const databasesDetailSummaryRoute = createRoute({
   getParentRoute: () => databasesDetailRoute,
   path: 'summary',
 }).lazy(() =>
-  import('./databasesLazyRoutes').then((m) => m.databaseDetailLazyRoute)
+  import('src/features/Databases/DatabaseDetail/databaseDetailLazyRoute').then(
+    (m) => m.databaseDetailLazyRoute
+  )
 );
 
 const databasesDetailBackupsRoute = createRoute({
   getParentRoute: () => databasesDetailRoute,
   path: 'backups',
 }).lazy(() =>
-  import('./databasesLazyRoutes').then((m) => m.databaseDetailLazyRoute)
+  import('src/features/Databases/DatabaseDetail/databaseDetailLazyRoute').then(
+    (m) => m.databaseDetailLazyRoute
+  )
 );
 
 const databasesDetailResizeRoute = createRoute({
   getParentRoute: () => databasesDetailRoute,
   path: 'resize',
 }).lazy(() =>
-  import('./databasesLazyRoutes').then((m) => m.databaseDetailLazyRoute)
+  import('src/features/Databases/DatabaseDetail/databaseDetailLazyRoute').then(
+    (m) => m.databaseDetailLazyRoute
+  )
 );
 
 const databasesDetailSettingsRoute = createRoute({
   getParentRoute: () => databasesDetailRoute,
   path: 'settings',
 }).lazy(() =>
-  import('./databasesLazyRoutes').then((m) => m.databaseDetailLazyRoute)
+  import('src/features/Databases/DatabaseDetail/databaseDetailLazyRoute').then(
+    (m) => m.databaseDetailLazyRoute
+  )
 );
 
 const databasesDetailConfigsRoute = createRoute({
   getParentRoute: () => databasesDetailRoute,
   path: 'configs',
 }).lazy(() =>
-  import('./databasesLazyRoutes').then((m) => m.databaseDetailLazyRoute)
+  import('src/features/Databases/DatabaseDetail/databaseDetailLazyRoute').then(
+    (m) => m.databaseDetailLazyRoute
+  )
 );
 
 const databasesDetailMetricsRoute = createRoute({
   getParentRoute: () => databasesDetailRoute,
   path: 'metrics',
 }).lazy(() =>
-  import('./databasesLazyRoutes').then((m) => m.databaseDetailLazyRoute)
+  import('src/features/Databases/DatabaseDetail/databaseDetailLazyRoute').then(
+    (m) => m.databaseDetailLazyRoute
+  )
 );
 
 const databasesDetailNetworkingRoute = createRoute({
   getParentRoute: () => databasesDetailRoute,
   path: 'networking',
 }).lazy(() =>
-  import('./databasesLazyRoutes').then((m) => m.databaseDetailLazyRoute)
+  import('src/features/Databases/DatabaseDetail/databaseDetailLazyRoute').then(
+    (m) => m.databaseDetailLazyRoute
+  )
 );
 
 export const databasesRouteTree = databasesRoute.addChildren([
